Update code & bug fixing and add some new security features i.e session token.

+++ admin/ajax/ajax_remove_mentee.php
@@ -5,16 +5,22 @@
     
     require ("../../partials/connection.php");
     
-    $rollNo = $_POST['rollNumber'];
-    
-    $sql = "DELETE FROM `mentee` WHERE `rollNo` = '$rollNo'";
-    
-    if (mysqli_query($conn, $sql)) {
+    session_name('admin_session');
+    session_start();
 
-        echo "remove mentee";
-    }
-    else {
+    if (isset($_SESSION['session_token'])) {
 
-        echo "failed to remove";
+        $rollNo = $_POST['rollNumber'];
+        
+        $sql = "DELETE FROM `mentee` WHERE `rollNo` = '$rollNo'";
+        
+        if (mysqli_query($conn, $sql)) {
+    
+            echo "remove mentee";
+        }
+        else {
+    
+            echo "failed to remove";
+        }
     }
 ?>
