+++ admin/ajax/ajax_remove_query.php
@@ -5,16 +5,22 @@
     
     require ("../../partials/connection.php");
     
-    $queryId = (int) $_POST['queryId'];
+    session_name('admin_session');
+    session_start();
+
+    if (isset($_SESSION['session_token'])) {
+
+        $queryId = (int) $_POST['queryId'];
+        
+        $sql = "DELETE FROM `queries` WHERE `id` = $queryId";
     
-    $sql = "DELETE FROM `queries` WHERE `id` = $queryId";
-
-    if (mysqli_query($conn, $sql)) {
-
-        echo "remove query";
-    }
-    else {
-
-        echo "failed to remove";
+        if (mysqli_query($conn, $sql)) {
+    
+            echo "remove query";
+        }
+        else {
+    
+            echo "failed to remove";
+        }
     }
 ?>
